Exercicio4 foi atualizado

+++ src/application/Exercicio4.java
@@ -11,18 +11,18 @@ public class Exercicio4 {
         Scanner sc = new Scanner(System.in);
 
         System.out.print("Número1: ");
-        int numero1 = sc.nextInt();
+        float numero1 = sc.nextFloat();
 
         System.out.print("Número2: ");
-        int numero2 = sc.nextInt();
+        float numero2 = sc.nextFloat();
 
         System.out.print("Número3: ");
-        int numero3 = sc.nextInt();
+        float numero3 = sc.nextFloat();
 
         System.out.print("Número4: ");
-        int numero4 = sc.nextInt();
+        float numero4 = sc.nextFloat();
 
-        int calculo = (numero1 * numero2) - (numero3 * numero4);
+        float calculo = (numero1 * numero2) - (numero3 * numero4);
 
         System.out.print("Diferença: " + calculo);
     }
